ugly solution for timedetla has problems with floats

+++ batch/extractFliesFromEntireDisk.py
@@ -285,7 +285,7 @@ for i in range(len(recRngs)):
             
             progress = curI / totI + 0.000000001
             passedTime = currentTime - startTime
-            eta = passedTime * (100.0 / progress)
+            eta = passedTime * int(100.0 / progress)
             finish = currentTime + eta
             
             status = \
@@ -316,7 +316,7 @@ for i in range(len(recRngs)):
             
             progress = curI / totI + 0.000000001
             passedTime = currentTime - startTime
-            eta = passedTime * (100.0 / progress)
+            eta = passedTime * int(100.0 / progress)
             finish = currentTime + eta
             
             status = \
